+++ src/PanelContent.jsx
@@ -4,27 +4,27 @@ import classnames from 'classnames';
 
 class PanelContent extends Component {
   shouldComponentUpdate(nextProps) {
-    return this.props.isActive || nextProps.isActive;
+    return this.props.forceRender || this.props.isActive || nextProps.isActive;
   }
 
   render() {
-    this._isActived = this._isActived || this.props.isActive;
+    this._isActived = this.props.forceRender || this._isActived || this.props.isActive;
     if (!this._isActived) {
       return null;
     }
-    const { prefixCls, isActive, children } = this.props;
+    const { prefixCls, isActive, children, destroyInactivePanel, forceRender, role } = this.props;
     const contentCls = classnames({
       [`${prefixCls}-content`]: true,
       [`${prefixCls}-content-active`]: isActive,
       [`${prefixCls}-content-inactive`]: !isActive,
     });
+    const child = !forceRender && !isActive && destroyInactivePanel ? null :
+      <div className={`${prefixCls}-content-box`}>{children}</div>;
     return (
       <div
         className={contentCls}
-        role="tabpanel"
-      >
-	{isActive?<div className={`${prefixCls}-content-box`}>{children}</div>: null}
-      </div>
+        role={role}
+      >{child}</div>
     );
   }
 }
@@ -33,6 +33,9 @@ PanelContent.propTypes = {
   prefixCls: PropTypes.string,
   isActive: PropTypes.bool,
   children: PropTypes.any,
+  destroyInactivePanel: PropTypes.bool,
+  forceRender: PropTypes.bool,
+  role: PropTypes.string,
 };
 
 export default PanelContent;