+++ tests/setup.js
@@ -1,0 +1,3 @@
+global.requestAnimationFrame = global.requestAnimationFrame || function requestAnimationFrame(cb) {
+  return setTimeout(cb, 0);
+};